+++ Ghost_in_the_Shell/client_folder_react_app/ethisim/src/components/EditorComponents/FlowDiagramComponents/HelperFunctions/node.js
@@ -1,0 +1,57 @@
+import React from 'react';
+import { Handle } from 'react-flow-renderer';
+
+export function actionNode({ data }) {
+    return (
+        <>
+            <Handle
+                type="target"
+                position="top"
+                onConnect={(params) => console.log('handle onConnect', params)}
+            />
+            {data.label}
+            <Handle
+                type="source"
+                position="bottom"
+                id="a"
+                style={{ left: '20%' }}
+            />
+            <Handle
+                type="source"
+                position="bottom"
+                id="b"
+                style={{ left: '80%' }}
+            />
+        </>
+    );
+}
+
+export function reflectionNode({ data }) {
+    return (
+        <>
+            <Handle type="target" position="top" />
+            {data.label}
+            <Handle type="source" position="bottom" />
+        </>
+    );
+}
+
+export function genericNode({ data }) {
+    return (
+        <>
+            <Handle type="target" position="top" />
+            {data.label}
+            <Handle type="source" position="bottom" />
+        </>
+    );
+}
+
+export function conversationNode({ data }) {
+    return (
+        <>
+            <Handle type="target" position="top" />
+            {data.label}
+            <Handle type="source" position="bottom" />
+        </>
+    );
+}
